+++ lib/generators/cms/component/profile_page/profile_page_generator.rb
@@ -6,84 +6,18 @@ module Cms
         include BasePaths
         include Actions
 
-        class_option :homepage_path,
-          :type => :string,
-          :default => nil,
-          :desc => 'Path to a CMS homepage, for which to create the contact form.'
-
-        class_option :skip_translation_import,
-          :type => :boolean,
-          :default => false,
-          :desc => 'Skip import of country translation files.'
+        class_option :cms_path,
+          type: :string,
+          default: nil,
+          desc: 'CMS parent path where the example profile should be placed under.',
+          banner: 'LOCATION'
 
         source_root File.expand_path('../templates', __FILE__)
-
-        def add_gems
-          gem('valid_email', '0.0.4')
-          gem('localized_country_select', '>= 0.9.2')
-
-          Bundler.with_clean_env do
-            run('bundle --quiet')
-          end
-        end
-
-        def import_translations
-          unless options[:skip_translation_import]
-            run('rake import:country_select LOCALE=en')
-            run('rake import:country_select LOCALE=de')
-          end
-        end
-
-        def extend_homepage
-          file = 'app/models/homepage.rb'
-          insert_point = "class Homepage < Obj\n"
-
-          data = []
-
-          data << '  include Cms::Attributes::ProfilePageLink'
-          data << ''
-
-          data = data.join("\n")
-
-          insert_into_file(file, data, :after => insert_point)
-        end
-
-        def extend_cell
-          file = 'app/cells/meta_navigation_cell.rb'
-          insert_point = "@current_user = current_user\n"
-
-          data = []
-
-          data << ''
-          data << '    @profile_page = page.homepage.profile_page'
-          data << ''
-
-          data = data.join("\n")
-
-          insert_into_file(file, data, :after => insert_point)
-        end
-
-        def extend_view
-          file = 'app/cells/meta_navigation/show.html.haml'
-          insert_point = "      = t('.meta')\n"
-
-          data = []
-
-          data << '    - if @current_user.logged_in?'
-          data << '      %li'
-          data << '        = link_to(cms_path(@profile_page)) do'
-          data << '          = display_title(@profile_page)'
-          data << ''
-
-          data = data.join("\n")
-
-          insert_into_file(file, data, :after => insert_point)
-        end
 
         def create_migration
           begin
             Model::ApiGenerator.new(behavior: behavior) do |model|
-              model.name = class_name
+              model.name = obj_class_name
               model.title = 'Page: Profile'
               model.attributes = [
                 {
@@ -91,20 +25,28 @@ module Cms
                   type: :boolean,
                   title: 'Show in navigation',
                 },
+                {
+                  name: sort_key_attribute_name,
+                  type: :string,
+                  title: 'Sort key',
+                },
               ]
             end
 
-            turn_model_into_page(class_name)
+            turn_model_into_page(obj_class_name)
           rescue Cms::Generators::DuplicateResourceError
           end
+        end
 
-          migration_template('example_migration.rb', 'cms/migrate/create_profile_page_example.rb')
+        def create_example
+          if example?
+            migration_template('example_migration.rb', 'cms/migrate/create_profile_page_example.rb')
+          end
         end
 
         def copy_app_directory
           directory('app', force: true)
           directory('config', force: true)
-          directory('spec', force: true)
         end
 
         def notice
@@ -115,20 +57,23 @@ module Cms
 
         private
 
-        alias_method :original_homepage_path, :homepage_path
-        def homepage_path
-          options[:homepage_path] || original_homepage_path
+        def example?
+          cms_path.present?
+        end
+
+        def cms_path
+          options[:cms_path]
         end
 
         def show_in_navigation_attribute_name
           'show_in_navigation'
         end
 
-        def profile_page_attribute_name
-          'profile_page_link'
+        def sort_key_attribute_name
+          'sort_key'
         end
 
-        def class_name
+        def obj_class_name
           'ProfilePage'
         end
       end