+++ lib/generators/cms/component/profile_page/templates/app/controllers/profile_page_controller.rb
@@ -1,6 +1,4 @@
 class ProfilePageController < CmsController
-  include RailsConnector::Crm::Localizable
-
   include Authorization
 
   def index