+++ lib/generators/cms/kickstart/templates/app/models/user_manager/authentication_failed.rb
@@ -1,3 +1,0 @@
-module UserManager
-  class AuthenticationFailed < StandardError; end
-end