+++ lib/generators/cms/kickstart/templates/config/initializers/user_manager.rb
@@ -1,3 +1,0 @@
-require 'user_manager/web_crm'
-
-Rails.application.config.user_manager = UserManager::WebCrm.new