+++ spec/generators/component/profile_page_generator_spec.rb
@@ -9,7 +9,7 @@ describe Cms::Generators::Component::ProfilePageGenerator do
   include GeneratorSpec::TestCase
 
   destination File.expand_path('../../../../tmp/generators', __FILE__)
-  arguments ['--skip_translation_import']
+  arguments ['--cms_path=/website/en']
 
   before(:all) do
     Cms::Generators::Attribute::ApiGenerator.send(:include, TestDestinationRoot)
@@ -23,20 +23,6 @@ describe Cms::Generators::Component::ProfilePageGenerator do
   end
 
   def prepare_environments
-    paths = {
-      models: "#{destination_root}/app/models",
-      cells: "#{destination_root}/app/cells",
-      meta_navigation: "#{destination_root}/app/cells/meta_navigation",
-    }
-
-    paths.each do |_, path|
-      mkdir_p(path)
-    end
-
-    File.open("#{destination_root}/Gemfile", 'w')
-    File.open("#{paths[:models]}/homepage.rb", 'w') { |f| f.write("class Homepage < Obj\n") }
-    File.open("#{paths[:cells]}/meta_navigation_cell.rb", 'w') { |f| f.write("    @current_user = current_user\n") }
-    File.open("#{paths[:meta_navigation]}/show.html.haml", 'w') { |f| f.write("      = t('.meta')\n") }
   end
 
   it 'creates files' do
@@ -44,10 +30,6 @@ describe Cms::Generators::Component::ProfilePageGenerator do
       directory 'app' do
         directory 'models' do
           file 'profile_page.rb'
-
-          file 'homepage.rb' do
-            contains 'include Cms::Attributes::ProfilePageLink'
-          end
         end
 
         directory 'presenters' do
@@ -61,19 +43,8 @@ describe Cms::Generators::Component::ProfilePageGenerator do
         directory 'concerns' do
           directory 'cms' do
             directory 'attributes' do
-              file 'profile_page_link.rb'
-            end
-          end
-        end
-
-        directory 'cells' do
-          file 'meta_navigation_cell.rb' do
-            contains '@profile_page = page.homepage.profile_page'
-          end
-
-          directory 'meta_navigation' do
-            file 'show.html.haml' do
-              contains '= display_title(@profile_page)'
+              file 'sort_key.rb'
+              file 'show_in_navigation.rb'
             end
           end
         end
@@ -91,17 +62,6 @@ describe Cms::Generators::Component::ProfilePageGenerator do
           migration 'create_profile_page'
           migration 'create_profile_page_example'
         end
-      end
-
-      directory 'spec' do
-        directory 'models' do
-          file 'profile_page_spec.rb'
-        end
-      end
-
-      file 'Gemfile' do
-        contains 'valid_email'
-        contains 'localized_country_select'
       end
     }
   end